requirements.txt
comtypes~=1.4.7
odfpy~=1.4.1
openpyxl~=3.1.5
pandas~=2.2.3
pypdf2~=3.0.1
pytesseract~=0.3.13
python-docx~=1.1.2
xlrd~=2.0.1
